requirements.txt
altgraph==0.17
autopep8==1.5.4
et-xmlfile==1.0.1
fbs==0.9.0
future==0.18.2
jdcal==1.4.1
macholib==1.14
openpyxl==3.0.5
pefile==2019.4.18
pycodestyle==2.6.0
PyInstaller==3.4
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
toml==0.10.2
